Improved ServerOptions Test case to determine default; run-tests script added
