+++ run-tests.sh
@@ -1,5 +1,5 @@
 # Creator:   VPR
-# Created:   February 1st, 2022
+# Created:   February 16th, 2022
 # Updated:   February 16th, 2022
 
 set -e pipefail
@@ -7,23 +7,35 @@ set -e errexit
 set -e nounset
 set -e xtrace
 
+# Run without setting environment variables
+if [ -f ./Server/Bin/run-tests ];
+then
+    ./Server/Bin/run-tests
+else
+    printf "'run-tests' not found.\n"
+    exit
+fi
+
+export SET=1
+
 export HOST="127.0.0.1"
 export PORT=8080
 
 export IPV4=0
-export TIMEOUT=30
 export MAX_CONNECTIONS=127
 export MAX_THREADS=8
 export N_THREADS=4
+export TIMEOUT=30
 export MEMORY_LIMIT=32768
 
 export MEM_CERT="${HOME}/certs/server_ca/certs/smoothstack_server.crt"
 export MEM_KEY="${HOME}/certs/server_ca/private/smoothstack_server.key"
 
-if [ -f ./Bin/utopia-server.exe ];
+# Run with environment variables set
+if [ -f ./Server/Bin/run-tests ];
 then
-    ./Bin/utopia-server.exe
+    ./Server/Bin/run-tests
 else
-    printf "'server.exe' not found.\n"
+    printf "'run-tests' not found.\n"
     exit
 fi
